--- src/config/db
@@ -1,37 +1,0 @@
-schema: wt
-
-
-
-table: diary
-
-+--------------+------------+------+-----+---------+----------------+
-| Field        | Type       | Null | Key | Default | Extra          |
-+--------------+------------+------+-----+---------+----------------+
-| id           | int(11)    | NO   | PRI | NULL    | auto_increment |
-| classes      | varchar(1) | YES  |     | NULL    |                |
-| created_date | datetime   | YES  |     | NULL    |                |
-| content      | text       | YES  |     | NULL    |                |
-| question     | text       | YES  |     | NULL    |                |
-| score        | int(1)     | YES  |     | NULL    |                |
-| user_id      | int(11)    | NO   | MUL | NULL    |                |
-+--------------+------------+------+-----+---------+----------------+
-
-
-
-
-
-table: user
-
-+--------------+--------------+------+-----+---------+----------------+
-| Field        | Type         | Null | Key | Default | Extra          |
-+--------------+--------------+------+-----+---------+----------------+
-| id           | int(11)      | NO   | PRI | NULL    | auto_increment |
-| email        | varchar(255) | NO   | UNI | NULL    |                |
-| pw           | varchar(45)  | NO   |     | NULL    |                |
-| kakaotoken   | varchar(255) | YES  |     | NULL    |                |
-| nickname     | varchar(255) | NO   |     | NULL    |                |
-| birthdate    | date         | NO   |     | NULL    |                |
-| sex          | varchar(1)   | NO   |     | NULL    |                |
-| address      | varchar(255) | NO   |     | NULL    |                |
-| created_date | datetime     | NO   |     | NULL    |                |
-+--------------+--------------+------+-----+---------+----------------+
