+++ src/routers/diary/index.js
@@ -1,11 +1,11 @@
 const express = require('express');
-const diaryCtrl = require('./diary.ctrl');
+const diaryCtrl = require('./diary.ctrl.js');
 const diaryRouter = express.Router();
 
-diaryRouter.post('/create', diaryCtrl.create);
-diaryRouter.get('/list', diaryCtrl.list);
-diaryRouter.get('/read/:id', diaryCtrl.read);
-diaryRouter.post('/update/:id', diaryCtrl.update);
+diaryRouter.post('/daily/:id', diaryCtrl.daily);
+diaryRouter.post('/monthly_mode_A/:id', diaryCtrl.monthly_mode_A);
+diaryRouter.post('/monthly_mode_B/:id', diaryCtrl.monthly_mode_B);
+diaryRouter.post('/annualy/:id', diaryCtrl.annualy);
 diaryRouter.delete('/delete/:id', diaryCtrl.delete);
 
 module.exports = diaryRouter;
