+++ src/test_data_generater.js
@@ -1,4 +1,4 @@
-const db = require('../config/db.js');
+const db = require('./config/db.js');
 const tools = require('./tools.js');
